+++ CHANGELOG.md
@@ -4,6 +4,7 @@ WordPress Dashboard Messages
 1.1.3
 -----
  - Upgrade dashicons
+ - Add Color scheme black
  - Improve Dashboard stylesheets for WP >= 5.5
  - Load legacy stylesheets for WP < 5.5